kernel::mem::slab: Supplement tests to Cache::allocate_object (#11)
In the existing implementation, Cache::allocate_object is tested indirectly through the tests 
for its implementing methods, which are private. The consequence is quite subtle. This PR 
addresses this issue by supplementing the tests for Cache::allocate_object itself. Additionally, 
Cache::allocate_object now returns a NoSlabAvailable error instead of AllocateFromNullSlab 
when there are no slabs or only full slabs.

#11