Introduce verify_cache_invariants_v2
The existing implementation didn't include the verification of the
expected slabs and allocated objects in the verification of the cache
invariants, but verified them separately. This commit aims to merge those
two verifications, and a v2 version of the verify_cache_invariants is
introduced. This function will eventually be renamed to replace the
existing verify_cache_invariants when all usages of the old function
are clear.