Enhance safe_slab_size
The existing implementation did not check if the total_slots is greater
than zero, nor did it compare the result with the Cache::min_slab_size,
which is amended by this commit. It also adds a reminder that the actual
total slot determined by the returned size may larger than the give value.