Remove allocate_from_empty_when_there_are_partial test
This is a subtle test case that the current implicit policy
allocates object from partial slabs first, so this test case will
never occur. We have decided to remove this test case because, even
if it becomes possible in the future, it should be tested thorugh
the tests for Cache::allocate_object.