kernel::mem::slab: Refactor Cache::allocate_from_empty (#15)
The existing implementation treated allocate_from_empty as a public function. 
This commit repositions it as a helper function that removes the sanitizing checks 
and the tests, relying instead on the checks and tests for the main function 
Cache::allocate_object. What remains is the cleaning of the 'EXCEPTION SAFETY' 
comment, which will be revisited after reviewing the relevant helper functions.